+++ src/sender-provider/index.ts
@@ -1,4 +1,4 @@
 export * from './baseSenderProvider.js';
 export * from './jitoSenderProvider.js';
 export * from './baseSenderProvider.js';
-export * from './jitoTipAccounts.js';
+export * from './jitoTips.js';
