+++ src/sender-provider/jitoBrowser.ts
@@ -1,11 +1,10 @@
 import { JitoJsonRpcClient } from 'jito-js-rpc';
 import { JitoClient, UUID } from './jitoTypes';
 import {
-  Connection,
   PublicKey,
   SystemProgram,
-  TransactionMessage,
-  VersionedTransaction
+  VersionedTransaction,
+  TransactionInstruction,
 } from '@solana/web3.js';
 import { getRandomTipAccount } from './jitoTipAccounts';
 
@@ -35,31 +34,7 @@ export const createBrowserClient = async (url: string): Promise<JitoClient> => {
       console.log(result);
     },
 
-    // NOTE: Probably want to use the transaction builder to build this
-    // to so we aren't passing the connection into this function
-    //
-    // A tip transaction is ~220 bytes (rough estimation 64 bytes for signature + 1 for length prefix)
-    // A tip instruction is ~150 bytes (exactly 152 bytes)
-    createTipTransaction: async (
-      connection: Connection,
-      payer: PublicKey,
-      tipAmount: number
-    ): Promise<VersionedTransaction> => {
-      const { blockhash } = await connection.getLatestBlockhash();
 
-      const tipIx = SystemProgram.transfer({
-        fromPubkey: payer,
-        toPubkey: getRandomTipAccount(),
-        lamports: tipAmount
-      });
-
-      const messageV0 = new TransactionMessage({
-        payerKey: payer,
-        recentBlockhash: blockhash,
-        instructions: [tipIx]
-      }).compileToV0Message();
-
-      return new VersionedTransaction(messageV0);
     }
   };
 };
